Remove regex-based codification, renumber scripts
Annotations are now all processed through the RAG pipeline.
- Move rules.yaml, 2_split_rules.py, and regex utils (load_rules,
  apply_rules, predict_code, normalize) to old/
- Rename scripts: 1_create_vector_db → 0_create_vector_db,
  3_run_rag → 2_run_rag
- Update Argo DAG: remove split-rules step, run-rag now depends
  directly on [create-vector-db, prune-annotations]
- Update README to reflect full pipeline
