+++ README.md
@@ -1,4 +1,4 @@
-# RAG classifier for product 
+# RAG classifier for product
 
 Using coicop classification.
 
@@ -10,34 +10,39 @@ uv sync
 
 ## Workflow
 
-### 1_create_vector_db.py
+Le pipeline est orchestré via Argo Workflows (`argo/pipeline.yaml`) selon le DAG suivant :
 
-Embed coicop's notices into a Qdrant vector database, using several different strategies.
+```
+prune-coicop ──→ create-vector-db ──┐
+             └──→ prune-annotations ──┘──→ run-rag
+```
 
-1. **Embedding Generation**:
-   - Text notices are processed through the VLLM embedding model
-   - Embeddings are generated using the `VLLM_EMBEDDING_URL` endpoint
-   - Authentication is handled via `VLLM_EMBEDDING_API_KEY`
+### 0_prunning_coicop.py
 
-2. **Vector Storage**:
-   - Generated embeddings are stored in Qdrant vector database
-   - Connection is established using `QDRANT_URL`, `QDRANT_API_KEY` and `QDRANT_API_PORT`
+Élague les hiérarchies linéaires de la nomenclature COICOP brute et exporte les notices filtrées ainsi qu'une table de correspondance vers S3.
+
+- Supprime le niveau 5 (Poste) de la nomenclature
+- Produit les notices prunées et la table de mapping niveau 4
+
+### 0_create_vector_db.py
+
+Encode les notices COICOP dans une base vectorielle Qdrant (plusieurs stratégies d'indexation).
+
+1. **Génération des embeddings** : les notices sont encodées via le modèle VLLM (`VLLM_EMBEDDING_URL`, `VLLM_EMBEDDING_API_KEY`)
+2. **Stockage vectoriel** : les embeddings sont indexés dans Qdrant (`QDRANT_URL`, `QDRANT_API_KEY`, `QDRANT_API_PORT`)
+
+### 1_prune_annotations.py
+
+Tronque les codes d'annotation au niveau 4 et applique la table de correspondance COICOP pour normaliser les codes de vérité terrain.
+
+- Dépend de `0_prunning_coicop.py` (requiert la table de mapping sur S3)
 
 ### 2_run_rag.py
 
-3. **Prompt Management**:
-   - Prompt templates are stored and managed in Langfuse
-   - Connection is established using `LANGFUSE_BASE_URL`, `LANGFUSE_PUBLIC_KEY` and `LANGFUSE_SECRET_KEY`
+Classifie les annotations via le pipeline RAG.
 
-4. **Retrieval and Generation**:
-   - Relevant context is retrieved from Qdrant based on query embeddings
-   - Retrieved context is passed to the VLLM generation model
-   - Final response is generated using the `VLLM_GENERATION_URL` endpoint
-   - Authentication is handled via `VLLM_GENERATION_API_KEY`
-
-5. **MLflow Logging**:
-   - Model performance metrics are logged to MLflow
-   - Connection is established using `MLFLOW_TRACKING_URI`
-
-
-
+3. **Gestion des prompts** : les templates sont stockés dans Langfuse (`LANGFUSE_BASE_URL`, `LANGFUSE_PUBLIC_KEY`, `LANGFUSE_SECRET_KEY`)
+4. **Retrieval et génération** :
+   - Les contextes pertinents sont récupérés depuis Qdrant
+   - La génération finale utilise le modèle VLLM (`VLLM_GENERATION_URL`, `VLLM_GENERATION_API_KEY`)
+5. **Logging MLflow** : les métriques sont enregistrées dans MLflow (`MLFLOW_TRACKING_URI`)
