+++ config/config.yaml
@@ -1,10 +1,10 @@
 # config.yaml
 
 annotations:
-  s3_path: "s3://projet-budget-famille/data/output-annotation-consolidated/annotations-consolidated-2026-03-26/raw_test.parquet"
-  s3_path_pruned: "s3://projet-budget-famille/data/workflow_outputs/prunning/raw_test_pruned.parquet"
+  s3_path: "s3://projet-budget-famille/data/regex/output/raw_test_without_regex.parquet"
+  s3_path_pruned: "s3://projet-budget-famille/data/workflow_outputs/prunning/raw_test_without_regex_pruned.parquet"
   s3_path_regex: "s3://projet-budget-famille/data/workflow_outputs/outputs_regex/raw_test_pruned_regex.parquet"
-  s3_path_rag: "s3://projet-budget-famille/data/workflow_outputs/outputs_regex/raw_test_pruned_noregex.parquet"
+  s3_path_rag: "s3://projet-budget-famille/data/workflow_outputs/outputs_regex/raw_test_without_regex_pruned_noregex.parquet"
   sample_size: null
   features:
     - id
