+++ old/2_split_rules.py
