+++ scripts/0_create_vector_db.py
